fix: refact StoreFlow to correct errors
Now the distinctUntilKeyChanged functions are passed other functions that with the help of the 'just-compare' library return only when the states are differents (didn't know 'distinctUntilKeyChanged' use a simple '===' eval.

The problem was caused by the SucursalesDetailsComponent.ResetButInfo whom was called more times that the really necessary making the 'InfiniteLoopFromHell'.

We've also removed the 'setTimeout' wraping the calls to the answers in 'Load[Open/Close]Answer' functions 'cause it wasn't really necessary for now at least.