refact: use ChartCard as an independent identity
We came back to use the chart-card.component as an independent identity of Card and not as its child. This is because it fixed the problems on iPads that the chart wasn't displaying its full height. Now ChartCard is a Copy, again, of the Card Component. We'll dig into it that later